+++ examples/examples_test.go
@@ -3,9 +3,12 @@
 package examples
 
 import (
+	"io/ioutil"
+	"net/http"
 	"os"
 	"path"
 	"testing"
+	"time"
 
 	"github.com/stretchr/testify/assert"
 
@@ -34,7 +37,16 @@ func TestExamples(t *testing.T) {
 
 	shortTests := []integration.ProgramTestOptions{
 		base.With(integration.ProgramTestOptions{Dir: path.Join(cwd, "minimal")}),
+		base.With(integration.ProgramTestOptions{Dir: path.Join(cwd, "topic")}),
+		base.With(integration.ProgramTestOptions{Dir: path.Join(cwd, "blob")}),
+		base.With(integration.ProgramTestOptions{Dir: path.Join(cwd, "queue")}),
 		base.With(integration.ProgramTestOptions{Dir: path.Join(cwd, "webserver")}),
+		base.With(integration.ProgramTestOptions{
+			Dir: path.Join(cwd, "http"),
+			ExtraRuntimeValidation: validateAPITest(func(body string) {
+				assert.Equal(t, body, "Hello World!")
+			}),
+		}),
 	}
 
 	longTests := []integration.ProgramTestOptions{
@@ -62,4 +74,27 @@ func TestExamples(t *testing.T) {
 
 func createEditDir(dir string) integration.EditDir {
 	return integration.EditDir{Dir: dir, ExtraRuntimeValidation: nil}
+}
+
+func validateAPITest(isValid func(body string)) func(t *testing.T, stack integration.RuntimeValidationStackInfo) {
+	return func(t *testing.T, stack integration.RuntimeValidationStackInfo) {
+		var resp *http.Response
+		var err error
+		url := stack.Outputs["url"].(string)
+		// Retry a couple times on 5xx
+		for i := 0; i < 5; i++ {
+			resp, err = http.Get(url)
+			if !assert.NoError(t, err) {
+				return
+			}
+			if resp.StatusCode < 500 {
+				break
+			}
+			time.Sleep(10 * time.Second)
+		}
+		defer resp.Body.Close()
+		body, err := ioutil.ReadAll(resp.Body)
+		assert.NoError(t, err)
+		isValid(string(body))
+	}
 }