+++ examples/http/index.ts
@@ -1,0 +1,15 @@
+import * as azure from '@pulumi/azure';
+
+const resourceGroup = new azure.core.ResourceGroup('example', { location: 'West US 2' });
+
+const greeting = new azure.appservice.HttpEventSubscription('greeting', {
+  resourceGroup,
+  callback: async (context, args) => {
+    return {
+      status: 200,
+      body: `Hello World!`
+    };
+  }
+});
+
+export const url = greeting.url;