+++ sdk/nodejs/eventhub/index.ts
@@ -20,3 +20,4 @@ export * from "./subscription";
 export * from "./subscriptionRule";
 export * from "./topic";
 export * from "./topicAuthorizationRule";
+export * from "./zMixins";